pyproject.toml
[tool.poetry] # https://python-poetry.org/docs/pyproject/, https://python-poetry.org/docs/basic-usage/
name = "app"
version = "0.1.0"
description = ""
authors = ["Me"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.1"
flask-sqlalchemy = "^3.1.1"
flask-bcrypt = "^1.0.1"
flask-login = "^0.6.3"
flask-marshmallow = "^1.1.0"
marshmallow-sqlalchemy = "^1.0.0"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
tqdm = "^4.66.1"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-env = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# https://stackoverflow.com/questions/49028611/pytest-cannot-find-module
[tool.pytest.ini_options]
pythonpath = [
  "."
]
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"

[tool.pytest_env]
FLASK_ENV = "testing"
SECRET_KEY = "test secret key"